app/requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
google-auth==2.28.2
google-oauth==1.0.1
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
rsa==4.9
scheduler==0.8.5
six==1.16.0
typeguard==4.1.5
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
